+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCSensitivityLabel/MSFT_SCSensitivityLabel.psm1
@@ -1795,7 +1795,12 @@ function Convert-StringToAdvancedSettings
                 Key   = $settingKey
                 Value = $values.Trim()
             }
-            $settings += $entry
+
+            # Only export the entry if it has a value
+            if ([String]::IsNullOrEmpty($entry.Value) -eq $false)
+            {
+                $settings += $entry
+            }
         }
     }
     return $settings